creating tables

+++ task_2.sql
@@ -21,7 +21,7 @@ CREATE TABLE Customers (
 
 CREATE TABLE Orders (
     order_id INT PRIMARY KEY,
-    customer_id BIGINT,
+    customer_id INT,
     order_date DATE,
     FOREIGN KEY (customer_id) REFERENCES Customers (customer_id)
 );
